% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_to_binary.R
\name{count_to_binary}
\alias{count_to_binary}
\title{Count to Binary Dataframe}
\usage{
count_to_binary(dat, x, y, time)
}
\arguments{
\item{dat}{A dataframe in person-period/long format.}

\item{x}{A numeric vector representing the total number of events (n).}

\item{y}{A numeric vector representing the total number of successes (k).}

\item{time}{A numeric vector representing the time points.}
}
\description{
This function receives the inputs of a count-based dataframe - typically longitudinal in nature - and
converts these counts to binary form where each row represents an independent event resulting in a 0 or 1 outcome.
This may be useful in situations of n events, choose k successes.
}
\examples{
new_dat = count_to_binary(dat, dat$x, dat$y, dat$time)

}
\keyword{Binary}
\keyword{Count}
\keyword{to}
