% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lm_output_table.R
\name{create_lm_output_table}
\alias{create_lm_output_table}
\title{This function is designed to extract the key information from an LM object and compile it into a tidy dataframe
for use in R Markdown reports. Information includes coefficient estimates, standard errors,
Z statistics, and p-values.}
\usage{
create_lm_output_table(mod_lm, dat)
}
\arguments{
\item{dat}{This is the dataframe used in fitting the lm object.}

\item{mod_glm}{This is an object of class 'lm'.}
}
\description{
This function is designed to extract the key information from an LM object and compile it into a tidy dataframe
for use in R Markdown reports. Information includes coefficient estimates, standard errors,
Z statistics, and p-values.
}
\examples{
dat = data.frame(x = rnorm(100, 0, 1),
y = 1.2 * x)

mod_lm = lm(y ~ x, data = dat)

lm_output = create_lm_output_table(mod_lm, dat)

}
\keyword{Output}
\keyword{Tidy}
