% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_table1.R
\name{df_table1}
\alias{df_table1}
\title{Adding Degrees of Freedom to table1}
\usage{
df_table1(x, ...)
}
\arguments{
\item{x}{This input is not necessary if used as an added function in the extra.cols argument in table1.
The input will automatically be detected by the function.}
}
\description{
This function allows you to add a column calculating degrees of freedom to the table1 function. The calculation is based on
a one-way ANOVA for continuous variables and chi-squared for categorical variables.
}
\examples{
table1::table1(~ . | group, data = dat, extra.cols = c(`DF`= df_table1))
}
\keyword{Degrees}
\keyword{freedom}
\keyword{of}
