% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsize_table1.R
\name{esize_table1}
\alias{esize_table1}
\title{Adding Effect Sizes to table1}
\usage{
esize_table1(x, ...)
}
\arguments{
\item{x}{This input is not necessary if used as an added function in the extra.cols argument in table1.
The input will automatically be detected by the function.}
}
\description{
This function allows you to add a column calculating effect sizes to the table1 function. The calculation is based on
eta-squared for continuous variables and Cramer's phi for categorical variables.
}
\examples{
table1::table1(~ . | group, data = dat, extra.cols = c(`Effect Size`= esize))

}
\keyword{Effect}
\keyword{Size}
