% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsqip_clean_data.R
\name{nsqip_clean_data}
\alias{nsqip_clean_data}
\title{Cleaning the raw National Surgical Quality Improvement Program (NSQIP) Dataset}
\usage{
nsqip_clean_data(dataset = dat)
}
\arguments{
\item{dataset}{the filtered and merged NSQIP data frame outputted by the "nsqip_subset_pufs" function}
}
\description{
This function is designed to perform routine data cleaning routines when using the raw NSQIP data.
Cleaning functions include missing data recoding, variable categorization/dichotomization, converting string
entries to usable formats, and changing variable classes. This function outputs a cleaned dataframe and variable summary.
}
\details{
NOTE: This function contains the sub-function update_label() to change variable labels.
}
