% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsqip_subset_pufs.R
\name{nsqip_subset_pufs}
\alias{nsqip_subset_pufs}
\title{Importing the raw National Surgical Quality Improvement Program (NSQIP) Dataset}
\usage{
nsqip_subset_pufs(
  puf_folder_path = "Raw Files",
  puf_years = 2012:2020,
  cpt_codes = NULL,
  cpt_codes_primary_only = TRUE,
  icd10_codes = NULL,
  icd9_codes = NULL,
  icd_codes_primary_only = TRUE,
  surgical_specialty = NULL
)
}
\arguments{
\item{puf_folder_path}{a text string representing the path of the folder containing the PUF text files; the text string must not end with "/"!}

\item{puf_years}{the years to include in the extract}

\item{cpt_codes}{the CPT codes to include in the extract. If NULL, then return all codes.
Based on the following fields: "CPT", "OTHERCPT1" - "OTHERCPT10", and "CONCPT1" - "CONCPT10"}

\item{cpt_codes_primary_only}{if TRUE, then only consider cases where one of the "cpt_codes" is the primary CPT code (the "CPT" field).
Ignore if cpt_codes = NULL.}

\item{icd10_codes}{a character vector of ICD-10 codes to subset the extract to. Ignore when NULL.
Based on the following fields: "PODIAG10", "PODIAG_OTHER10", "REOPOR1ICD101", "REOPOR2ICD101", "READMRELICD101", "READMUNRELICD101",
                               "READMRELICD102", "READMUNRELICD102", "READMRELICD103", "READMUNRELICD103", "READMRELICD104", "READMUNRELICD104",
                               "READMRELICD105", "READMUNRELICD105"}

\item{icd9_codes}{a character vector of ICD-9 codes to subset the extract to (these become rare in more recent years). Ignore when NULL.
Based on the following fields: "PODIAG", "PODIAG_OTHER", "REOPORICD91", "REOPOR2ICD91", "READMRELICD91", "READMUNRELICD91", "READMRELICD92",
                               "READMUNRELICD92", "READMRELICD93", "READMUNRELICD93", "READMRELICD94", "READMUNRELICD94", "READMRELICD95",
                               "READMUNRELICD95"}

\item{icd_codes_primary_only}{if TRUE, then only use the "PODIAG10" and "PODIAG" fields for filtering.}

\item{surgical_specialty}{a character vector of surgical specialty to subset to (based on the "SURGSPEC" field).}
}
\description{
This function is designed to specify a directory with multiple NSQIP datasets to be filtered and merged.
Due to the size of many NSQIP datasets, importing each is likely to cause out-of-memory errors on most personal computers.
This function is intended to be run prior to running the nsqip_clean_data() function. Exports a merged dataframe object.
}
