% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_overall_and_stratified_tables.R
\name{Overall.And.Stratified}
\alias{Overall.And.Stratified}
\title{Creating Overall and Stratified Tables}
\usage{
Overall.And.Stratified(
  dataset = dat,
  strata.variable = NULL,
  variables,
  variable.labels = NULL,
  factor.variables = NULL,
  factors_reported.categorical.analyzed.ordinal = NULL,
  exact.variables = NULL,
  exact.variables_detect.and.add.to.list = TRUE,
  categorical.variables_force.exclusion.from.exact.analysis = NULL,
  categorical.variables_report.all.levels = FALSE,
  non.normal.variables = NULL,
  all.patients.row.name = "All Patients",
  descriptors_include = TRUE,
  descriptor_categorical = ", N (\%)",
  descriptor_continuous_default = ", Mean (SD)",
  descriptor_continuous_nonnormal = ", Median [Q1,Q3]",
  digits_pvalue = 3,
  digits_categorical = 1,
  digits_continuous = 2,
  missing.data_include.as.list = TRUE,
  missing.data_include.as.separate.table = TRUE,
  missing.data_add.to.table = FALSE,
  missing.data_row.label = "Missing Obs.",
  missing.data_value.prefix = "N="
)
}
\arguments{
\item{dataset}{A data.frame object containing, at minimum, all of the variables contained in the table.}

\item{strata.variable}{= the name of the categorical/pordinal variable that you wish to stratify on. May also be left as NULL for an overall table.}

\item{variables}{= a vector of the variable names you wish to include as rows in the table.}

\item{variable.labels}{= an optional vector of labels that will be displayed instead of the variable name, Must be the same length as the variables vector.}

\item{factor.variables}{= an optional vector of variable names that are numeric but should be coereced into categorical.}

\item{factors_reported.categorical.analyzed.ordinal}{= an optional vector of variable names where frequencies are reported but the variables are analyzed as ordinal.}

\item{exact.variables}{= an optional vector of variable names where Fisher's exact should be used.}

\item{exact.variables_detect.and.add.to.list}{= TRUE/FALSE whether or not to run a chi-square test on the categorical variables and if a warning is produced run Fisher's exact.}

\item{categorical.variables_force.exclusion.from.exact.analysis}{= an optional vector of variable names that will be coerced into the chi-square test.}

\item{categorical.variables_report.all.levels}{= if TRUE, then both levels of two-level categorical variables will be reported}

\item{non.normal.variables}{= an optional vector of variable names where a non-parametric test will be used instead of a t-test/one-way ANOVA.}

\item{all.patients.row.name}{= a character string that will be displayed in the top row of the table which typically contrains the strata-specific frequencies.}

\item{descriptors_include}{= TRUE/FALSE to include the statistic description after the variable label in the rows of the table.}

\item{descriptor_categorical}{= a character string of the descriptor to include after categorical variables.}

\item{descriptor_continuous_default}{= a character string of the descriptor to include after continuous variables.}

\item{descriptor_continuous_nonnormal}{= a character string of the descriptor to include after non-normal continuous variables.}

\item{digits_pvalue}{= a numeric value for the number of digits to report for p-values.}

\item{digits_categorical}{= a numeric value for the number of digits to report for percentages.}

\item{digits_continuous}{= a numberic value for the number of digits to report for numeric values.}

\item{missing.data_include.as.list}{= TRUE/FALSE for whether or not to generate a character string to describe the variables with missing data.}

\item{missing.data_include.as.separate.table}{= TRUE/FALSE for whether or not to generate a separate table summarizing the variables with missing data.}

\item{missing.data_add.to.table}{= TRUE/FALSE whether or not to add an additional row below a variable for the purposes of reporting missing values.}

\item{missing.data_row.label}{=  a character string for the row label that will be used for missing data rows}

\item{missing.data_value.prefix}{= a character string for the pre-fix to include before the missing data frequencies in the body of the table.}
}
\description{
This function is designed to create a summary table using lists of continuous and categorical variables (specified).
Overall or stratified tables can be constructed or if a stratification variable is included. Additionally,
p-values are output using either ANOVA (for continuous variables) or Chi-square (for categorical variables) tests.
}
\keyword{Summary}
\keyword{Table}
