% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue_table1.R
\name{pvalue_table1}
\alias{pvalue_table1}
\title{Adding P-values to table1}
\usage{
pvalue_table1(x, ...)
}
\arguments{
\item{x}{This input is not necessary if used as an added function in the extra.cols argument in table1.
The input will automatically be detected by the function.}
}
\description{
This function allows you to add a column calculating effect sizes to the table1 function. The calculation is based on
one-way ANOVA for continuous variables and chi-squared for categorical variables.
}
\examples{
data("sleep")

table1::table1(~ extra | group, data = sleep, extra.cols = c(`P-Value`= pvalue_table1))

}
\keyword{P-value;}
\keyword{Significance}
\keyword{test}
