% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_residuals.R
\name{quantile_residuals}
\alias{quantile_residuals}
\title{This function is used to compute the quantile residuals in a fitted beta regression model.}
\usage{
quantile_residuals(betareg_mod, dat, outcome)
}
\arguments{
\item{betareg_mod}{The fitted beta regression model using the `betareg` function.}

\item{dat}{The dataframe used to fit the `betareg` model also used in this function. The dimensions of
the dataframe must match those of the dataframe used to fit the `betareg` object.}

\item{outcome}{The response variable from the original dataframe used in the `betareg` object.
Specified as a character string.}
}
\description{
When fitting beta regression models, the quantile residuals are argued to be the more reasonable and useful measure of
deviation from fitted and actual values in the model (Pereira, 2019). This stands in contrast to the currently
more typically used adjusted weighted standardized residual 1 or the weighted standardized residual 2.
}
\details{
Currently, this function only works for models fitted using the `betareg` function in `R`.
}
\examples{
data("GasolineYield", package = "betareg")

gy = betareg(yield ~ batch + temp | temp, data = GasolineYield)

summary(gy)

quantile_residuals(gy, GasolineYield, "yield")

}
\references{
Pereira, G. H. (2019). On quantile residuals in beta regression,
Communications in Statistics - Simulation and Computation 48(1), 302-316,
doi: 10.1080/03610918.2017.1381740.
}
\keyword{Beta}
\keyword{Quantile}
\keyword{regression,}
\keyword{residuals}
